fix(multimodal): merge compat shim into package — remove shadowed multimodal_processor.py (#3554)
multimodal_processor.py was unreachable because Python always resolves the
multimodal_processor/ package first. Merge ModalInput alias, MultiModalProcessor
compat class, and multimodal_processor singleton into package __init__.py so
existing importers get the same symbols without any changes. Delete the
shadowed file.
